Intermediate commit trying to work through using a high resolution timer for the CapacitySensor

+++ Capacitor_Noise_Display/Capacitor_Display/Capacitor_Display.ino
@@ -27,7 +27,7 @@ void setup()
 
 void loop()
 {
-   uint32_t chargeTime = 0;
+   float chargeTime = 0;
    while (sensor.tryDequeue(chargeTime))
    {
       chargeTime10.set(chargeTime);
